Moved step test utils into its own module
All of the test utility types and functions all required a lot of `cfg()` declarations that were becoming unwieldy.  So this refactors them into its own module to cut down on a lot of the boilerplate.
