+++ mmids-core/src/workflows/steps/workflow_forwarder/tests.rs
@@ -2,7 +2,8 @@ use super::*;
 use crate::test_utils;
 use crate::workflows::definitions::WorkflowStepType;
 use crate::workflows::metadata::MediaPayloadMetadataCollection;
-use crate::workflows::steps::{FuturesChannelInnerResult, StepTestContext};
+use crate::workflows::steps::futures_channel::FuturesChannelInnerResult;
+use crate::workflows::steps::test_utils::StepTestContext;
 use crate::workflows::MediaType;
 use anyhow::{anyhow, Result};
 use bytes::{Bytes, BytesMut};
